GRILL 137 Fix SessionManager.stop()

+++ service/src/java/org/apache/hive/service/cli/session/SessionManager.java
@@ -122,7 +122,7 @@ public class SessionManager extends CompositeService {
     super.stop();
     if (backgroundOperationPool != null) {
       backgroundOperationPool.shutdown();
-      int timeout = hiveConf.getIntVar(ConfVars.HIVE_SERVER2_ASYNC_EXEC_SHUTDOWN_TIMEOUT);
+      long timeout = hiveConf.getLongVar(ConfVars.HIVE_SERVER2_ASYNC_EXEC_SHUTDOWN_TIMEOUT);
       try {
         backgroundOperationPool.awaitTermination(timeout, TimeUnit.SECONDS);
       } catch (InterruptedException e) {
@@ -346,3 +346,4 @@ public class SessionManager extends CompositeService {
     }
   }
 }
+
